requirements.txt
numpy
torch>=1.3
torchvision
numba==0.40.0
tensorboardX
easydict
pyyaml
permissive_dict
opencv-python
scipy
tqdm
